+++ packages/amplify-graphql-model-transformer/rds-lambda/clients/DBClient.ts
@@ -1,6 +1,31 @@
 import { Knex } from 'knex';
+import {Request} from '../interfaces/BaseRequest';
+import {ListRequest, SortDirection} from '../interfaces/ListRequest';
 
-export interface DBClient {
-    client: Knex;
-    getClient(): Promise<any | Knex<any, any[]>>;
+export abstract class DBClient {
+  client: Knex;
+  abstract getClient(): Promise<any | Knex<any, any[]>>;
+
+  protected addKeyConditions = (query: any, request: Request) => {
+    const keys = request.args.metadata.keys || [];
+    keys.forEach((key) => {
+      query.where(key, request.args.input.get(key));
+    });
+  }
+
+  protected addSortConditions = (query: any, request: ListRequest) => {
+    // order using sort keys
+    const sortDirection = request.args.sortDirection || SortDirection.ASC;
+    const keys = request.args.metadata.keys || [];
+    if(keys.length > 1) {
+      const sortKeys = request.args.metadata.keys.slice(1);
+      const orderByConditions = sortKeys.map((sortKey) => {
+        return {
+          column: sortKey,
+          order: sortDirection.toString().toLowerCase(),
+        };
+      });
+      query.orderBy(orderByConditions);
+    }
+  }
 }
