+++ packages/amplify-graphql-model-transformer/rds-lambda/package.json
@@ -10,7 +10,7 @@
     "babel-jest": "^29.1.2",
     "jest": "^29.1.2",
     "knex": "^2.3.0",
-    "mysql": "^2.18.1",
+    "mysql2": "^3.1.0",
     "pg": "^8.8.0",
     "ts-jest": "^29.0.3",
     "uuid": "^9.0.0"
